requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
psutil>=5.9.0
pika>=1.3.2
Pillow>=9.0.0
requests>=2.28.0
python-multipart>=0.0.5
dotenv>=0.9.9
